moved the tables to a separate module
